Rem FOE (speed issues), improved optiflow
Field of Expansion drops 5fps, which implies a heavy overhead for the pi.

However, having a vectorised version of the motion detection allows for a couple of other algorithms which might prove fruitful.

Removed some of the noise in the vectorised motion images via anding with the scalar motion image.